--- src/core/ui/themes/default/components/MuiTooltip.ts
@@ -1,11 +1,0 @@
-import { Components, Theme } from '@mui/material/styles';
-
-const MuiTooltip = (_theme: Theme): Components['MuiTooltip'] => {
-  return {
-    defaultProps: {
-      arrow: true,
-    },
-  };
-};
-
-export default MuiTooltip;
