--- src/domain/community/application/views/PendingApplications/PendingApplicationsView.tsx
@@ -1,79 +1,0 @@
-import React, { FC } from 'react';
-import { useTranslation } from 'react-i18next';
-import Box from '@mui/material/Box';
-import Card from '../../../../../common/components/core/Card';
-import { Loading } from '../../../../../common/components/core';
-import { ApplicationWithType } from '../../../../../common/utils/application/getApplicationWithType';
-import PendingApplicationContainer from '../../containers/PendingApplicationContainer';
-import PendingApplicationView from './PendingApplicationView';
-
-export interface PendingApplicationsViewEntities {
-  applications: ApplicationWithType[];
-}
-
-export interface PendingApplicationsViewActions {
-  handleDelete: (application: ApplicationWithType) => void;
-}
-
-export interface PendingApplicationsViewState {
-  isDeleting: boolean;
-  loading: boolean;
-}
-
-export interface PendingApplicationsViewOptions {
-  canEdit: boolean;
-}
-
-export interface PendingApplicationsViewProps {
-  entities: PendingApplicationsViewEntities;
-  actions: PendingApplicationsViewActions;
-  state: PendingApplicationsViewState;
-  options: PendingApplicationsViewOptions;
-}
-
-const PendingApplicationsView: FC<PendingApplicationsViewProps> = ({
-  entities: containerEntities,
-  actions: containerActions,
-  state: containerState,
-  options: containerOptions,
-}) => {
-  const { t } = useTranslation();
-
-  return (
-    <Box marginY={1}>
-      <Card primaryTextProps={{ text: t('pages.user-profile.applications.title') }}>
-        {containerState.loading && <Loading />}
-        {containerEntities.applications.map((x, i) => (
-          // todo: refactor to use only one container; remove PendingApplicationContainer
-          <PendingApplicationContainer key={i} entities={{ application: x }}>
-            {(entities, actions, state) => (
-              <PendingApplicationView
-                key={i}
-                entities={{
-                  application: x,
-                  applicationDetails: entities.applicationDetails,
-                  typeName: entities.typeName,
-                  url: entities.url,
-                }}
-                actions={{
-                  handleDelete: containerActions.handleDelete,
-                  handleDialogOpen: actions.handleDialogOpen,
-                  handleDialogClose: actions.handleDialogClose,
-                }}
-                state={{
-                  loading: containerState.loading || state.loading,
-                  isDeleting: containerState.isDeleting,
-                  isDialogOpened: state.isDialogOpened,
-                  loadingDialog: state.loadingDialog,
-                }}
-                options={containerOptions}
-              />
-            )}
-          </PendingApplicationContainer>
-        ))}
-      </Card>
-    </Box>
-  );
-};
-
-export default PendingApplicationsView;
