--- src/domain/community/contributor/organization/pages/CardProps.ts
@@ -1,4 +1,0 @@
-export default interface CardProps {
-  id: string;
-  spaceID?: string;
-}
