--- src/domain/platform/admin/components/Common/VisualSegment.ts
@@ -1,7 +1,0 @@
-import * as yup from 'yup';
-
-export const visualSegmentSchema = yup.object().shape({
-  avatar: yup.string(),
-  background: yup.string(),
-  banner: yup.string(),
-});
