--- src/domain/shared/components/ActivityLog/index.ts
@@ -1,1 +1,0 @@
-export * from './ActivityComponent';
