--- src/domain/shared/components/Backdrops/AuthenticationBackdrop.tsx
@@ -1,30 +1,0 @@
-import React, { FC } from 'react';
-import { useTranslation } from 'react-i18next';
-import { useNavigate } from 'react-router-dom';
-import BackdropWithMessage, { BackdropProps } from './BackdropWithMessage';
-import { _AUTH_LOGIN_PATH } from '../../../../core/auth/authentication/constants/authentication.constants';
-import WrapperButton from '../../../../common/components/core/WrapperButton';
-import { useAuthenticationContext } from '../../../../core/auth/authentication/hooks/useAuthenticationContext';
-
-const AuthenticationBackdrop: FC<BackdropProps> = ({ children, blockName }) => {
-  const { t } = useTranslation();
-  const navigate = useNavigate();
-
-  const { isAuthenticated } = useAuthenticationContext();
-
-  return (
-    <BackdropWithMessage
-      message={t('components.backdrop.authentication', { blockName })}
-      children={children}
-      show={!isAuthenticated}
-      template={
-        <WrapperButton
-          onClick={() => navigate(_AUTH_LOGIN_PATH, { replace: true })}
-          text={t('authentication.sign-in')}
-        />
-      }
-    />
-  );
-};
-
-export default AuthenticationBackdrop;
