+++ src/domain/shared/components/ClampedTypography/ClampedTypography.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { styled, Typography, TypographyProps } from '@mui/material';
 
-export interface ClampedTypographyProps extends TypographyProps {
+interface ClampedTypographyProps extends TypographyProps {
   clamp: number;
   component?: React.ElementType;
 }
