--- src/domain/shared/components/ClampedTypography/index.ts
@@ -1,2 +1,0 @@
-export { default as ClampedTypography } from './ClampedTypography';
-export * from './ClampedTypography';
