--- src/domain/shared/components/Text/Constants.ts
@@ -1,5 +1,0 @@
-import { Theme } from '@mui/material/styles';
-
-export const LINE_HEIGHT = 4;
-
-export const getLineHeight = (theme: Theme) => theme.spacing(LINE_HEIGHT);
