+++ src/domain/shared/components/search-cards/SearchChallengeCard.tsx
@@ -1,8 +1,8 @@
 import React, { FC } from 'react';
 import SearchBaseJourneyCard, { SearchBaseJourneyCardProps } from './base/SearchBaseJourneyCard';
 
-export type SearchChallengeCardImplProps = Omit<SearchBaseJourneyCardProps, 'journeyTypeName'>;
-export interface SearchChallengeCardProps extends SearchChallengeCardImplProps {}
+type SearchChallengeCardImplProps = Omit<SearchBaseJourneyCardProps, 'journeyTypeName'>;
+interface SearchChallengeCardProps extends SearchChallengeCardImplProps {}
 
 export const SearchChallengeCard: FC<SearchChallengeCardProps> = ({ children, ...rest }) => {
   return <SearchBaseJourneyCard journeyTypeName="challenge" {...rest} />;
